DB setting

+++ Models/AppointmentResponse.cs
@@ -10,7 +10,7 @@ namespace GP2.Models
     {
         [Key]
         [Required]
-        public long AppointmentId { get; set; }
+        public int AppointmentId { get; set; }
 
         [Required(ErrorMessage = " Please enter a valid group name.")]
         public string Name { get; set; }
